+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- Added `kamu reset` command that reverts the dataset back to the specified state
+### Added
+- New `kamu reset` command that reverts the dataset back to the specified state
+### Fixed
+- Improved the `kamu log` command's performance on datasets with high block counts and introruced a `--limit` parameter
 
 ## [0.94.0] - 2022-06-22
 ### Added